Add is_maine to Forum and Reply Serializers Add suggested_tags To Forum Serializer When is_mine == True Delete Tag From Reply Serializer Delete my_vote map From vote_data map if is_my_vote_exist == False

+++ vote/views.py
@@ -19,10 +19,18 @@ class VotedItemViewSet(ModelViewSet):
 
     def destroy(self, request, *args, **kwargs):
         instance = self.get_object()
-        vote.send_robust(self.__class__, old_instance=instance,
-                         new_instance=None)
-        instance.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        if request.user.id == instance.user.id:
+            vote.send_robust(self.__class__, old_instance=instance,
+                             new_instance=None)
+            instance.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        return Response(status=status.HTTP_401_UNAUTHORIZED)
+
+    def update(self, request, *args, **kwargs):
+        instance = self.get_object()
+        if request.user.id == instance.user.id:
+            return super().update(request, *args, **kwargs)
+        return Response(status=status.HTTP_401_UNAUTHORIZED)
 
 
 @api_view(['GET'])
